Refactored code; allows mouse/keyboard separately enabled
